Fixed warning on windows.
git-svn-id: https://svn.opendtect.org/od/trunk@27010 a985855d-d384-4b45-ae33-31a410998c3c

+++ include/Basic/valseries.h
@@ -486,7 +486,7 @@ RT MultiArrayValueSeries<RT,AT>::value( od_int64 idx ) const
 	return RT();
 
     idx -= arridx*chunksize_;
-    return  ptrs_[arridx][idx];
+    return  ptrs_[(int)arridx][(int)idx];
 }
 
 
@@ -505,7 +505,7 @@ void MultiArrayValueSeries<RT,AT>::setValue( od_int64 idx, RT v )
 	return;
 
     idx -= arridx*chunksize_;
-    ptrs_[arridx][idx] = v;
+    ptrs_[(int) arridx][(int) idx] = v;
 }
